+++ internal/service/thirdpartyintegration/resource_third_party_integration_migration_test.go
@@ -1,0 +1,11 @@
+package thirdpartyintegration_test
+
+import (
+	"testing"
+
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/testutil/mig"
+)
+
+func TestMigThirdPartyIntegration_basic(t *testing.T) {
+	mig.CreateAndRunTest(t, basicTestCase(t))
+}
